pyproject.toml
[project]
name = "fastapi-archetype"
version = "0.0.2"
description = "A reference FastAPI application demonstrating enterprise patterns"
readme = "README.md"
authors = [
    { name = "Tommaso", email = "tommaso-meledina@users.noreply.github.com" }
]
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic-settings>=2.13.1",
    "PyJWT[crypto]>=2.9.0",
    "slowapi>=0.1.9",
    "structlog>=24.0.0",
    "pymysql>=1.1.2",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
    "aiosqlite>=0.22.1",
    "aiomysql>=0.3.2",
    "greenlet>=3.3.2",
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipdb",
    "nest-asyncio>=1.6.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "ruff>=0.15.4",
    "ty",
]

[tool.ruff]
target-version = "py314"
src = ["src"]
line-length = 88

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "Depends"]

[tool.ruff.lint.isort]
known-first-party = ["fastapi_archetype"]

[tool.ty.environment]
python-version = "3.14"

[tool.ty.src]
include = ["src", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--strict-markers -ra"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
omit = [
    "*/implementations/mock_dummy_service.py",
]
